+++ actfw/core/capture.py
