+++ actfw/core/v4l2/types.py
